feat(hook): add useDonationButton.js hook  Added `useDonationButton.js` hook for handling donation button logic.  refactor(footer): remove functionality, include boilerplate footer  Removed previous functionality from `footer.jsx` and replaced it with a boilerplate footer.  chore(main): add strict mode to React application  Enabled strict mode in the React application within `main.jsx`.  fix(app): correct page rendering issues  Fixed rendering issues in `app.jsx` to ensure proper page rendering.  fix(graphql): resolve undefined error in typeDefs.js  Updated `typeDefs.js` to fix undefined errors related to GraphQL schema definitions.
feat(hook): add useDonationButton.js hook

Added `useDonationButton.js` hook for handling donation button logic.

refactor(footer): remove functionality, include boilerplate footer

Removed previous functionality from `footer.jsx` and replaced it with a boilerplate footer.

chore(main): add strict mode to React application

Enabled strict mode in the React application within `main.jsx`.

fix(app): correct page rendering issues

Fixed rendering issues in `app.jsx` to ensure proper page rendering.

fix(graphql): resolve undefined error in typeDefs.js

Updated `typeDefs.js` to fix undefined errors related to GraphQL schema definitions.
